+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -93,5 +93,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-06-23T12:22:11.975Z")
+    created: Optional[datetime] = parse_datetime("2025-06-24T05:41:49.240Z")
     r"""The date and time the resource was created"""