+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatapointop.py
@@ -1098,5 +1098,5 @@ class UpdateDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-06-23T12:22:08.975Z")
+    updated: Optional[datetime] = parse_datetime("2025-06-24T05:41:46.101Z")
     r"""The date and time the resource was last updated"""