+++ packages/orq-rc/RELEASES.md
@@ -1799,3 +1799,13 @@ Based on:
 - [python v3.9.0-rc.1] packages/orq-rc
 ### Releases
 - [PyPI v3.9.0-rc.1] https://pypi.org/project/orq-ai-sdk/3.9.0-rc.1 - packages/orq-rc
+
+## 2025-06-24 05:43:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.568.2 (2.634.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.9.0-rc.2] packages/orq-rc
+### Releases
+- [PyPI v3.9.0-rc.2] https://pypi.org/project/orq-ai-sdk/3.9.0-rc.2 - packages/orq-rc